Macros expander WIP
