+++ packages/taquito-michel-codec/src/index.ts
@@ -1,4 +1,4 @@
-export * from "./ast";
-export * from "./parse";
-export * from "./emitter";
-export * from "./validate";
+export * from "./micheline";
+export * from "./micheline-parser";
+export * from "./micheline-emitter";
+export * from "./michelson-validator";
