+++ packages/taquito-michel-codec/src/macros.ts
@@ -1,0 +1,5 @@
+import { Prim } from "./micheline";
+
+export function expandMacros(ex: Prim): Prim | Prim[] {
+    return ex;
+}
